+++ src/main/java/webserver/http/Controller.java
@@ -1,0 +1,5 @@
+package webserver.http;
+
+public interface Controller {
+    void execute(HttpRequest httpRequest, HttpResponse httpResponse);
+}
