+++ src/main/java/webserver/http/RegexpMapping.java
@@ -1,15 +1,15 @@
 package webserver.http;
 
-import webserver.http.controller.Controller;
+import webserver.http.dispatcher.HttpRequestMapping;
 
 import java.util.regex.Pattern;
 
-class RegexpMapping implements HttpRequestMapping {
+public class RegexpMapping implements HttpRequestMapping {
     private Pattern uriPattern;
     private HttpMethod method;
     private final Controller controller;
 
-    RegexpMapping(String uriPatternRegexp, HttpMethod method, Controller controller) {
+    public RegexpMapping(String uriPatternRegexp, HttpMethod method, Controller controller) {
         this.uriPattern = Pattern.compile(uriPatternRegexp);
         this.method = method;
         this.controller = controller;
