--- src/main/java/webserver/http/controller/Controller.java
@@ -1,8 +1,0 @@
-package webserver.http.controller;
-
-import webserver.http.HttpRequest;
-import webserver.http.HttpResponse;
-
-public interface Controller {
-    void execute(HttpRequest httpRequest, HttpResponse httpResponse);
-}
